# doc/api/swagger.yaml
basePath: /api/v2
definitions:
...
  database.Widget:
    properties:
      customProperties:
        description: Custom properties of widget as JSON string
        type: string
      dashboardID:
        description: ID of dashboard to which widget belongs
        type: integer
      height:
        description: Height of widget
        type: integer
      id:
        type: integer
      isLocked:
        description: Locked state of widget
        type: boolean
      minHeight:
        description: Minimal height of widget
        type: integer
      minWidth:
        description: Minimal width of widget
        type: integer
      name:
        description: Name of widget
        type: string
      type:
        description: Type of widget
        type: string
      width:
        description: Width of widget
        type: integer
      x:
        description: X position of widget
        type: integer
      "y":
        description: Y position of widget
        type: integer
      z:
        description: Z position of widget
        type: integer
    type: object
...
  docs.ResponseError:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  docs.ResponseWidget:
    properties:
      widget:
        $ref: '#/definitions/database.Widget'
        type: object
    type: object
  docs.ResponseWidgets:
    properties:
      widgets:
        items:
          $ref: '#/definitions/database.Widget'
        type: array
    type: object
...
  widget.addWidgetRequest:
    properties:
      widget:
        $ref: '#/definitions/widget.validNewWidget'
        type: object
    type: object
  widget.updateWidgetRequest:
    properties:
      widget:
        $ref: '#/definitions/widget.validUpdatedWidget'
        type: object
    type: object
  widget.validNewWidget:
    properties:
      CustomProperties:
        type: string
      DashboardID:
        type: integer
      Height:
        type: integer
      IsLocked:
        type: boolean
      MinHeight:
        type: integer
      MinWidth:
        type: integer
      Name:
        type: string
      Type:
        type: string
      Width:
        type: integer
      X:
        type: integer
      "Y":
        type: integer
      Z:
        type: integer
    required:
    - DashboardID
    - Height
    - Name
    - Type
    - Width
    type: object
  widget.validUpdatedWidget:
    properties:
      CustomProperties:
        type: string
      Height:
        type: integer
      IsLocked:
        type: boolean
      MinHeight:
        type: integer
      MinWidth:
        type: integer
      Name:
        type: string
      Type:
        type: string
      Width:
        type: integer
      X:
        type: integer
      "Y":
        type: integer
      Z:
        type: integer
    type: object
host: localhost:4000
info:
  contact:
    email: sonja.happ@eonerc.rwth-aachen.de
    name: Sonja Happ
  description: |-
    This is the VILLASweb Backend API v2.0.
    Parts of this API are still in development. Please check https://git.rwth-aachen.de/acs/public/villas/web-backend-go for more information.
  license:
    name: GNU GPL 3.0
    url: http://www.gnu.de/documents/gpl-3.0.en.html
  title: VILLASweb Backend API
  version: "2.0"
paths:
...
  /widgets:
    get:
      operationId: getWidgets
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Dashboard ID
        in: query
        name: dashboardID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Widgets to which belong to dashboard
          schema:
            $ref: '#/definitions/docs.ResponseWidgets'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get all widgets of dashboard
      tags:
      - widgets
    post:
      consumes:
      - application/json
      operationId: addWidget
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Widget to be added incl. ID of dashboard
        in: body
        name: inputWidget
        required: true
        schema:
          $ref: '#/definitions/widget.addWidgetRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Widget that was added
          schema:
            $ref: '#/definitions/docs.ResponseWidget'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Add a widget to a dashboard
      tags:
      - widgets
  /widgets/{widgetID}:
    delete:
      operationId: deleteWidget
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Widget ID
        in: path
        name: widgetID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Widget that was deleted
          schema:
            $ref: '#/definitions/docs.ResponseWidget'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Delete a widget
      tags:
      - widgets
    get:
      operationId: getWidget
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Widget ID
        in: path
        name: widgetID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Widget that was requested
          schema:
            $ref: '#/definitions/docs.ResponseWidget'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Get a widget
      tags:
      - widgets
    put:
      consumes:
      - application/json
      operationId: updateWidget
      parameters:
      - description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Widget to be updated
        in: body
        name: inputWidget
        required: true
        schema:
          $ref: '#/definitions/widget.updateWidgetRequest'
          type: object
      - description: Widget ID
        in: path
        name: widgetID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Widget that was updated
          schema:
            $ref: '#/definitions/docs.ResponseWidget'
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "422":
          description: Unprocessable entity
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/docs.ResponseError'
            type: object
      summary: Update a widget
      tags:
      - widgets
swagger: "2.0"